+++ libs/shared/src/lib/components/widgets/grid-settings/button-config/button-config.component.ts
@@ -355,9 +355,20 @@ export class ButtonConfigComponent
     this.modificationsArray.push(
       this.fb.group({
         field: ['', Validators.required],
-        value: ['', Validators.required],
+        value: [''],
       })
     );
+  }
+
+  /**
+   * Set the value of a modification to null
+   *
+   * @param index The index of the modification
+   */
+  public setModificationValueToNull(index: number): void {
+    const modifications = this.modificationsArray.value;
+    modifications[index].value = null;
+    this.modificationsArray.setValue(modifications);
   }
 
   /**