+++ libs/shared/src/lib/components/widgets/map-settings/edit-layer-modal/layer-aggregation/layer-aggregation.component.ts
@@ -1,12 +1,6 @@
-import {
-  AfterViewInit,
-  Component,
-  Input,
-  ViewChild,
-  ViewContainerRef,
-} from '@angular/core';
+import { DomPortal } from '@angular/cdk/portal';
+import { Component, Input } from '@angular/core';
 import { FormGroup } from '@angular/forms';
-import { BehaviorSubject, Subject, takeUntil } from 'rxjs';
 
 /** Available reduction types */
 const AVAILABLE_REDUCTION_TYPES = ['cluster'];
@@ -19,29 +13,11 @@ const AVAILABLE_REDUCTION_TYPES = ['cluster'];
   templateUrl: './layer-aggregation.component.html',
   styleUrls: ['./layer-aggregation.component.scss'],
 })
-export class LayerAggregationComponent implements AfterViewInit {
+export class LayerAggregationComponent {
+  /** Current form group */
   @Input() formGroup!: FormGroup;
+  /** Map dom portal */
+  @Input() mapPortal?: DomPortal;
+  /** Available reduction types */
   public reductionTypes = AVAILABLE_REDUCTION_TYPES;
-
-  // Display of map
-  @Input() currentMapContainerRef!: BehaviorSubject<ViewContainerRef | null>;
-  @ViewChild('mapContainer', { read: ViewContainerRef })
-  mapContainerRef!: ViewContainerRef;
-  @Input() destroyTab$!: Subject<boolean>;
-
-  ngAfterViewInit(): void {
-    this.currentMapContainerRef
-      .pipe(takeUntil(this.destroyTab$))
-      .subscribe((viewContainerRef) => {
-        if (viewContainerRef) {
-          if (viewContainerRef !== this.mapContainerRef) {
-            const view = viewContainerRef.detach();
-            if (view) {
-              this.mapContainerRef.insert(view);
-              this.currentMapContainerRef.next(this.mapContainerRef);
-            }
-          }
-        }
-      });
-  }
 }