+++ libs/shared/src/lib/components/widgets/map-settings/edit-layer-modal/layer-datasource/layer-datasource.component.ts
@@ -1,20 +1,16 @@
 import {
-  AfterViewInit,
-  ChangeDetectorRef,
   Component,
   EventEmitter,
   Input,
   OnInit,
   Output,
   ViewChild,
-  ViewContainerRef,
 } from '@angular/core';
 import { Apollo, QueryRef } from 'apollo-angular';
 import {
   takeUntil,
   BehaviorSubject,
   Observable,
-  Subject,
   distinctUntilChanged,
 } from 'rxjs';
 import {
@@ -47,6 +43,7 @@ import { MapLayersService } from '../../../../../services/map/map-layers.service
 import { Fields } from '../../../../../models/layer.model';
 import { GraphQLSelectComponent } from '@oort-front/ui';
 import { Dialog } from '@angular/cdk/dialog';
+import { DomPortal } from '@angular/cdk/portal';
 
 /** Default items per resources query, for pagination */
 const ITEMS_PER_PAGE = 10;
@@ -80,37 +77,39 @@ const ADMIN_FIELDS = [
 })
 export class LayerDatasourceComponent
   extends UnsubscribeComponent
-  implements OnInit, AfterViewInit
+  implements OnInit
 {
+  /** Current form group */
   @Input() formGroup!: FormGroup;
+  /** Query to get resource */
   @Input() resourceQuery!: BehaviorSubject<ResourceQueryResponse | null>;
-  public origin!: FormControl<string | null>;
-
-  // Resource
-  public resourcesQuery!: QueryRef<ResourcesQueryResponse>;
-  public resource: Resource | null = null;
+  /** Available fields */
+  @Input() fields$!: Observable<Fields[]>;
+  /** Map dom portal */
+  @Input() mapPortal?: DomPortal;
+  /** Emit new fields */
+  @Output() fields: EventEmitter<Fields[]> = new EventEmitter<Fields[]>();
+  /** Reference to resource graphql select */
   @ViewChild(GraphQLSelectComponent)
   resourceSelect?: GraphQLSelectComponent;
-
-  // Reference data
-  public refDatasQuery!: QueryRef<ReferenceDatasQueryResponse>;
-  public refData: ReferenceData | null = null;
+  /** Reference to reference data graphql select */
   @ViewChild(GraphQLSelectComponent)
   refDataSelect?: GraphQLSelectComponent;
-
-  // Aggregation and layout
+  /** Type of origin ( resource or reference data ) */
+  public origin!: FormControl<string | null>;
+  /** Query to get resources */
+  public resourcesQuery!: QueryRef<ResourcesQueryResponse>;
+  /** Selected resource */
+  public resource: Resource | null = null;
+  /** Query to get reference data */
+  public refDatasQuery!: QueryRef<ReferenceDatasQueryResponse>;
+  /** Selected reference data */
+  public refData: ReferenceData | null = null;
+  /** Selected aggregation */
   public aggregation: Aggregation | null = null;
+  /** Selected layout */
   public layout: Layout | null = null;
-
-  @Input() fields$!: Observable<Fields[]>;
-  @Output() fields: EventEmitter<Fields[]> = new EventEmitter<Fields[]>();
-
-  // Display of map
-  @Input() currentMapContainerRef!: BehaviorSubject<ViewContainerRef | null>;
-  @ViewChild('mapContainer', { read: ViewContainerRef })
-  mapContainerRef!: ViewContainerRef;
-  @Input() destroyTab$!: Subject<boolean>;
-
+  /** Admin fields */
   public adminFields = ADMIN_FIELDS;
 
   /**
@@ -121,15 +120,13 @@ export class LayerDatasourceComponent
    * @param gridLayoutService Shared layout service
    * @param aggregationService Shared aggregation service
    * @param mapLayersService Shared map layer Service.
-   * @param cdr Change detector
    */
   constructor(
     private apollo: Apollo,
     private dialog: Dialog,
     private gridLayoutService: GridLayoutService,
     private aggregationService: AggregationService,
-    private mapLayersService: MapLayersService,
-    private cdr: ChangeDetectorRef
+    private mapLayersService: MapLayersService
   ) {
     super();
   }
@@ -250,23 +247,6 @@ export class LayerDatasourceComponent
           this.fields.next(
             this.getFieldsFromRefData(this.refData.fields || [])
           );
-        }
-      });
-  }
-
-  ngAfterViewInit(): void {
-    this.currentMapContainerRef
-      .pipe(takeUntil(this.destroyTab$))
-      .subscribe((viewContainerRef) => {
-        if (viewContainerRef) {
-          if (viewContainerRef !== this.mapContainerRef) {
-            const view = viewContainerRef.detach();
-            if (view) {
-              this.mapContainerRef.insert(view);
-              this.cdr.detectChanges();
-              this.currentMapContainerRef.next(this.mapContainerRef);
-            }
-          }
         }
       });
   }